feat: add solution to cjsToEsm.mjs
